requirements/base.txt
DALL-E
dlib
keras
Pillow
requests
scipy
tensorflow
tqdm
